// src/app/models/enums/role-actions.enum.ts
export enum RoleActionTypeEnum {
  AddEdit = 'addEdit',
  View = 'view',
  Delete = 'delete',
  Run = 'run',
}
export enum RoleActionsPageNameEnum {
  Users = 'Users',
  UserRoles = 'UserRoles',
  CompanyConfiguration = 'CompanyConfiguration',
  AllCompanies = 'AllCompanies',
  AllUsers = 'AllUsers',
  WorkflowTemplates = 'WorkflowTemplates',
  AgentTemplates = 'AgentTemplates',
  SnapshotTemplate = 'SnapshotTemplates',
  Snap2Template = 'Snap2Templates',
  tbAuth = 'TailorbrandAuth',
  Insights = 'Insights',
  AgentExecution = 'AgentExecution',
  UtilityConfigurations = 'UtilityConfigurations',
  Dashboard = 'Dashboard',
  KioskTemplates = 'KioskTemplates',
  MonitorTemplates = 'MonitorTemplates',
  HOOTemplates = 'HOOTemplates',
  SchedulerTemplates = 'SchedulerTemplates',
  MobileTemplates = 'MobileTemplates',
  HomeInterfaces = 'HomeInterfaces',
  Branches = 'Branches',
  GeneralReports = 'GeneralReports',
  AuditReports = 'AuditReports',
  LogsAnalytics = 'LogsAnalytics',
  BranchGeneralConfiguration = 'BranchGeneralConfiguration',
  BranchContactInfo = 'BranchContactInfo',
  ContactInfo = 'Contact Info.',
  BranchWorkFlow = 'BranchWorkFlow',
  BranchKioskConfigurations = 'BranchKioskConfigurations',
  BranchMonitorConfigurations = 'BranchMonitorConfigurations',
  BranchDeskConfigurations = 'BranchDeskConfigurations',
  BranchMobileInterface = 'BranchMobileInterface',
  BranchAdditionalSetting = 'BranchAdditionalSetting',
  BranchAdvancedSetting = 'BranchAdvancedSetting',
  Help = 'Help',
  MonitorDiagnostics = "MonitorDiagnostics"
}